+++ packages/cf-deploy-config-writer/src/cf-writer/base-config.ts
@@ -2,7 +2,7 @@ import { join } from 'path';
 import { create as createStorage } from 'mem-fs';
 import { create, type Editor } from 'mem-fs-editor';
 import { type Logger } from '@sap-ux/logger';
-import { generateSupportingConfig } from '../utils';
+import { generateSupportingConfig, fileExists } from '../utils';
 import LoggerHelper from '../logger-helper';
 import { createMTA, validateMtaConfig, addRoutingConfig } from '../mta-config';
 import { type CFBaseConfig, type MTABaseConfig } from '../types';
@@ -25,7 +25,7 @@ export async function generateBaseConfig(config: CFBaseConfig, fs?: Editor, logg
     }
     logger?.debug(`Generate base configuration using: \n ${JSON.stringify(config)}`);
     validateMtaConfig(config);
-    if (fs.exists(join(config.mtaPath, config.mtaId))) {
+    if (fileExists(fs, join(config.mtaPath, config.mtaId))) {
         throw new Error(t('error.mtaFolderAlreadyExists'));
     }
     createMTA(config as MTABaseConfig);