+++ packages/fiori-app-sub-generator/test/int/fiori-elements/writing-v2.test.ts
@@ -153,7 +153,8 @@ describe('Generate v2 apps', () => {
 
         const testState: Partial<State> = {
             project: Object.assign({}, v2Project, {
-                name: testProjectName
+                name: testProjectName,
+                enableVirtualEndpoints: true
             }) as Project,
             floorplan: FloorplanFE.FE_LROP,
             service: Object.assign({}, v2Service, {
@@ -291,7 +292,7 @@ describe('Generate v2 apps', () => {
             floorplan: FloorplanFE.FE_OVP,
             service: ovpService,
             entityRelatedConfig: {
-                filterEntityType: {
+                filterEntitySet: {
                     entitySetName: 'GlobalFilters',
                     entitySetType: 'GlobalFilters'
                 }