+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.66
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.21
+
+## 0.4.65
+
+### Patch Changes
+
+-   Updated dependencies [294bbe3]
+    -   @sap-ux/project-access@1.29.20
+
+## 0.4.64
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.19
+
 ## 0.4.63
 
 ### Patch Changes