+++ packages/project-integrity/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/project-integrity
 
+## 0.1.20
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.21
+
+## 0.1.19
+
+### Patch Changes
+
+-   Updated dependencies [294bbe3]
+    -   @sap-ux/project-access@1.29.20
+
+## 0.1.18
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.19
+
 ## 0.1.17
 
 ### Patch Changes