+++ packages/repo-app-import-sub-generator/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @sap-ux/repo-app-download-sub-generator
 
+## 0.2.7
+
+### Patch Changes
+
+-   f01b502: Improve test coverage and add App Validation for @sap-ux/repo-app-import-sub-generator
+
+## 0.2.6
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-writer@0.0.106
+-   @sap-ux/fiori-elements-writer@2.3.2
+-   @sap-ux/launch-config@0.9.3
+-   @sap-ux/project-access@1.29.21
+-   @sap-ux/odata-service-inquirer@2.3.3
+-   @sap-ux/axios-extension@1.20.0
+-   @sap-ux/fiori-generator-shared@0.11.2
+-   @sap-ux/project-input-validator@0.5.4
+-   @sap-ux/inquirer-common@0.6.39
+
+## 0.2.5
+
+### Patch Changes
+
+-   Updated dependencies [294bbe3]
+    -   @sap-ux/project-access@1.29.20
+    -   @sap-ux/abap-deploy-config-writer@0.0.105
+    -   @sap-ux/axios-extension@1.20.0
+    -   @sap-ux/fiori-generator-shared@0.11.1
+    -   @sap-ux/launch-config@0.9.2
+    -   @sap-ux/odata-service-inquirer@2.3.2
+    -   @sap-ux/project-input-validator@0.5.3
+    -   @sap-ux/fiori-elements-writer@2.3.1
+    -   @sap-ux/inquirer-common@0.6.38
+
+## 0.2.4
+
+### Patch Changes
+
+-   Updated dependencies [1ca4004]
+    -   @sap-ux/fiori-generator-shared@0.11.0
+    -   @sap-ux/fiori-elements-writer@2.3.0
+    -   @sap-ux/odata-service-inquirer@2.3.1
+    -   @sap-ux/inquirer-common@0.6.37
+    -   @sap-ux/abap-deploy-config-writer@0.0.104
+    -   @sap-ux/launch-config@0.9.1
+    -   @sap-ux/project-access@1.29.19
+    -   @sap-ux/axios-extension@1.20.0
+    -   @sap-ux/project-input-validator@0.5.2
+
+## 0.2.3
+
+### Patch Changes
+
+-   Updated dependencies [afd26c8]
+    -   @sap-ux/odata-service-inquirer@2.3.0
+    -   @sap-ux/fiori-elements-writer@2.2.0
+
 ## 0.2.2
 
 ### Patch Changes